--- player_app/src/frame_history.rs
@@ -1,46 +1,0 @@
-use egui::util::History;
-
-pub struct FrameHistory {
-    frame_times: History<f32>,
-}
-
-impl Default for FrameHistory {
-    fn default() -> Self {
-        let max_age: f32 = 1.0;
-        let max_len = (max_age * 300.0).round() as usize;
-        Self {
-            frame_times: History::new(0..max_len, max_age),
-        }
-    }
-}
-
-impl FrameHistory {
-    // Called first
-    pub fn on_new_frame(&mut self, now: f64, previous_frame_time: Option<f32>) {
-        let previous_frame_time = previous_frame_time.unwrap_or_default();
-        if let Some(latest) = self.frame_times.latest_mut() {
-            *latest = previous_frame_time; // rewrite history now that we know
-        }
-        self.frame_times.add(now, previous_frame_time); // projected
-    }
-
-    pub fn mean_frame_time(&self) -> f32 {
-        self.frame_times.average().unwrap_or_default()
-    }
-
-    pub fn fps(&self) -> f32 {
-        1.0 / self.frame_times.mean_time_interval().unwrap_or_default()
-    }
-
-    //TODO: Fix it
-    // pub fn limit_fps(&mut self, previous_render_time: std::time::Instant) {
-    //     let target_frametime = std::time::Duration::from_secs_f64(1.0 / 144.); // 1/60 seconde
-    //     let elapsed = previous_render_time.elapsed();
-
-    //     if elapsed < target_frametime {
-    //         println!("Sleeping for {:?} ({:?} - {:?})", target_frametime - elapsed, target_frametime, elapsed);
-    //         let delay = target_frametime - elapsed;
-    //         std::thread::sleep(delay);
-    //     }
-    // }
-}
