+++ src/common/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod config;
+pub mod ipc_message;
+pub mod preset;
+pub mod audio;
